requirements.txt
openai>=1.0.0
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
notion-client
python-dotenv
